+++ src/Fable.AST/Fable.fs
@@ -7,13 +7,15 @@ type EntityPath =
     | AssemblyPath of string
     /// Only used to reference entities in core assemblies without a path
     | CoreAssemblyName of string
+    | PrecompiledLib of sourcePath: string * assemblyPath: string
 
 type EntityRef =
     { FullName: string
       Path: EntityPath }
     member this.SourcePath =
         match this.Path with
-        | SourcePath p -> Some p
+        | SourcePath p
+        | PrecompiledLib(p,_) -> Some p
         | AssemblyPath _ | CoreAssemblyName _ -> None
 
 type DeclaredType =
@@ -97,6 +99,7 @@ type Entity =
     abstract IsValueType: bool
     abstract IsFSharpExceptionDeclaration: bool
     abstract IsInterface: bool
+    abstract IsMeasure: bool
 
 type Type =
     | Measure of fullname: string
@@ -366,7 +369,7 @@ type Expr =
     /// Lambdas are curried, they always have a single argument (which can be unit)
     | Lambda of arg: Ident * body: Expr * name: string option
     /// Delegates are uncurried functions, can have none or multiple arguments
-    | Delegate of args: Ident list * body: Expr * name: string option
+    | Delegate of args: Ident list * body: Expr * name: string option * isArrow: bool
     | ObjectExpr of members: MemberDecl list * typ: Type * baseCall: Expr option
 
     // Type cast and tests
@@ -427,7 +430,7 @@ type Expr =
         | IfThenElse (_, expr, _, _)
         | DecisionTree (expr, _) -> expr.Type
         | Lambda(arg, body, _) -> LambdaType(arg.Type, body.Type)
-        | Delegate(args, body, _) -> DelegateType(args |> List.map (fun a -> a.Type), body.Type)
+        | Delegate(args, body, _, _) -> DelegateType(args |> List.map (fun a -> a.Type), body.Type)
 
     member this.Range: SourceLocation option =
         match this with
@@ -438,7 +441,7 @@ type Expr =
         | DecisionTree _
         | DecisionTreeSuccess _ -> None
         | Lambda (_, e, _)
-        | Delegate (_, e, _)
+        | Delegate (_, e, _, _)
         | TypeCast (e, _) -> e.Range
         | IdentExpr id -> id.Range
         | Extended(_,r)