+++ src/Fable.Core/RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+### 3.3.1
+
+* JS.Decorator/ReflectedDecorator attributes
+* JS.Function
+
+### 3.2.9
+
+* Add `FormattableString.GetStrings()` extension
+
 ### 3.2.8
 
 * Experimental.namesofLambda